pyproject.toml
[project]
name = "mcp-fitbit-obsidian"
version = "0.1.0"
description = "MCP server to sync Fitbit health data to Obsidian daily notes"
requires-python = ">=3.10"
dependencies = [
    "mcp>=1.0.0",
    "requests>=2.31.0",
    "python-dotenv>=1.0.0",
    "cherrypy>=18.0.0",
]

[project.scripts]
run-server = "src.server:main"
auth = "scripts.auth:main"

[dependency-groups]
dev = [
    "pytest>=8.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src", "scripts"]

[tool.pytest.ini_options]
testpaths = ["tests"]
